Varnish now shuts down correctly
This changes start.sh to make containers shut down correctly when
`docker stop` is invoked.

Previous behaviour: docker stop is invoked, the process ignores it
and after the timeout docker kill is automatically invoked which
shuts down the process with a sigkill (not giving varnish a chance
to clean up properly)

Current behaviour: when docker stop is invoked varnish receives
a sigterm and exits properly.